+++ integration/rabbitmq/e2e/nack-and-requeue.e2e-spec.ts
@@ -60,7 +60,7 @@ describe('Nack and Requeue', () => {
     const moduleFixture = await Test.createTestingModule({
       providers: [SubscribeService],
       imports: [
-        RabbitMQModule.forRootAsync(RabbitMQModule, {
+        RabbitMQModule.forRootAsync({
           useFactory: () => ({
             exchanges: [
               {