+++ integration/rabbitmq/src/named-connection/named-connection.module.ts
@@ -11,7 +11,7 @@ const uri = `amqp://rabbitmq:rabbitmq@${rabbitHost}:${rabbitPort}`;
 
 @Module({
   imports: [
-    RabbitMQModule.forRootAsync(RabbitMQModule, {
+    RabbitMQModule.forRootAsync({
       useFactory: () => ({
         name: CONNECTION_NAME,
         exchanges: [