+++ packages/discovery/src/discovery.service.ts
@@ -4,8 +4,6 @@ import { PATH_METADATA } from '@nestjs/common/constants';
 import { STATIC_CONTEXT } from '@nestjs/core/injector/constants';
 import { InstanceWrapper } from '@nestjs/core/injector/instance-wrapper';
 import { Module } from '@nestjs/core/injector/module';
-import { ModulesContainer } from '@nestjs/core/injector/modules-container';
-import { MetadataScanner } from '@nestjs/core/metadata-scanner';
 import { flatMap, get, isNil, some, uniqBy } from 'lodash';
 import {
   DiscoveredClass,
@@ -14,6 +12,7 @@ import {
   Filter,
   MetaKey,
 } from './discovery.interfaces';
+import { MetadataScanner, ModulesContainer } from '@nestjs/core';
 
 /**
  * Attempts to retrieve meta information from a Nest DiscoveredClass component
@@ -223,7 +222,8 @@ export class DiscoveryService {
     return {
       name: wrapper.name as string,
       instance: instanceHost.instance,
-      injectType: wrapper.metatype,
+      // TODO: remove nullish coalescing operator to return undefined when dropping NestJS 10 support
+      injectType: wrapper.metatype ?? undefined,
       dependencyType: get(instanceHost, 'instance.constructor'),
       parentModule: {
         name: nestModule.metatype.name,