+++ packages/graphql-request/src/graphql-request-module-definition.ts
@@ -1,0 +1,7 @@
+import { ConfigurableModuleBuilder } from '@nestjs/common';
+import { GraphQLRequestModuleConfig } from './graphql-request-types';
+
+export const { ConfigurableModuleClass, MODULE_OPTIONS_TOKEN } =
+  new ConfigurableModuleBuilder<GraphQLRequestModuleConfig>()
+    .setClassMethodName('forRoot')
+    .build();