+++ packages/graphql-request/src/index.ts
@@ -1,3 +1,4 @@
 export * from './graphql-request.constants';
 export * from './graphql-request.decorators';
 export * from './graphql-request.module';
+export * from './graphql-request-types';