+++ packages/rabbitmq/src/rabbitmq-module-definition.ts
@@ -1,0 +1,7 @@
+import { ConfigurableModuleBuilder } from '@nestjs/common';
+import { RabbitMQConfig } from './rabbitmq.interfaces';
+
+export const { ConfigurableModuleClass, MODULE_OPTIONS_TOKEN } =
+  new ConfigurableModuleBuilder<RabbitMQConfig>()
+    .setClassMethodName('forRoot')
+    .build();