+++ packages/stripe/src/stripe-module-definition.ts
@@ -1,0 +1,7 @@
+import { ConfigurableModuleBuilder } from '@nestjs/common';
+import { StripeModuleConfig } from './stripe.interfaces';
+
+export const { ConfigurableModuleClass, MODULE_OPTIONS_TOKEN } =
+  new ConfigurableModuleBuilder<StripeModuleConfig>()
+    .setClassMethodName('forRoot')
+    .build();