+++ packages/webhooks/src/webhooks.decorators.ts
@@ -1,19 +1,7 @@
-import { makeInjectableDecorator } from '@golevelup/nestjs-common';
-import {
-  WEBHOOK_MODULE_CONFIG_TOKEN,
-  WEBHOOK_MODULE_PROVIDED_CONFIG_TOKEN,
-} from './webhooks.constants';
-
-/**
- * Injects the Webhook Config provided to this module (may have optional values)
- */
-export const InjectProvidedWebhookConfig = makeInjectableDecorator(
-  WEBHOOK_MODULE_PROVIDED_CONFIG_TOKEN
-);
+import { Inject } from '@nestjs/common';
+import { WEBHOOK_MODULE_CONFIG_TOKEN } from './webhooks.constants';
 
 /**
  * Injects the final Webhook Config (optional values filled in with defaults) from this module
  */
-export const InjectWebhookConfig = makeInjectableDecorator(
-  WEBHOOK_MODULE_CONFIG_TOKEN
-);
+export const InjectWebhookConfig = () => Inject(WEBHOOK_MODULE_CONFIG_TOKEN);