+++ packages/testing/ts-jest/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.2](https://github.com/golevelup/nestjs/compare/@golevelup/ts-jest@0.6.1...@golevelup/ts-jest@0.6.2) (2025-01-24)
+
+**Note:** Version bump only for package @golevelup/ts-jest
+
 ## [0.6.1](https://github.com/golevelup/nestjs/compare/@golevelup/ts-jest@0.6.0...@golevelup/ts-jest@0.6.1) (2024-11-12)
 
 ### Bug Fixes