+++ packages/testing/ts-vitest/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.2](https://github.com/golevelup/nestjs/compare/@golevelup/ts-vitest@0.5.1...@golevelup/ts-vitest@0.5.2) (2025-01-24)
+
+**Note:** Version bump only for package @golevelup/ts-vitest
+
 ## [0.5.1](https://github.com/golevelup/nestjs/compare/@golevelup/ts-vitest@0.5.0...@golevelup/ts-vitest@0.5.1) (2024-11-06)
 
 **Note:** Version bump only for package @golevelup/ts-vitest